:sparkles: feat: 참여 일자 int 에서 byte 로 자료형 변경 #146

+++ src/main/java/com/habitpay/habitpay/domain/challenge/domain/Challenge.java
@@ -44,7 +44,7 @@ public class Challenge extends BaseTime {
     private int numberOfParticipants;
 
     @Column(nullable = false)
-    private int participatingDays;
+    private byte participatingDays;
 
     @Column(nullable = false)
     private int feePerAbsence;
